refactor(peras): merge finalize layer files
