+++ fsmeta/runtime/peras/finalize_layers.go
@@ -5,6 +5,27 @@ package peras
 
 import "context"
 
+// sealedTrackingLayer finalizes the runtime read view for an installed
+// segment. Non-materialized segments are added to the sealed overlay; all
+// segments are removed from the visible overlay and counted in segment stats.
+type sealedTrackingLayer struct {
+	runtime *Runtime
+}
+
+func newSealedTrackingLayer(runtime *Runtime) SegmentFinalizeLayer {
+	if runtime == nil {
+		return nil
+	}
+	return &sealedTrackingLayer{runtime: runtime}
+}
+
+func (l *sealedTrackingLayer) FinalizeSegment(_ context.Context, req SegmentFinalizeRequest) error {
+	if l == nil || l.runtime == nil {
+		return ErrRuntimeInvalid
+	}
+	return l.runtime.installSegment(req.Plan, req.Segment, req.MaterializeMVCC)
+}
+
 // completionIndexLayer records each segment's per-operation completion in the
 // runtime's read state so duplicate-operation submits (SubmitVisible retries)
 // can short-circuit without re-issuing the work. The layer is in-memory only
